updated the following: - html     - tried to add js "mobileToDesktop" to header. still working on that..     - updated src, removing the initial "."     - created desktop-header section with minilogo     - updated on-sale class to include "on-sale" - css     - main         - removed unnecessary unique-section classes         - updated all px into rem         - created "mobile-menu-section"         - added align-content center to "mobile-menu"         - removed test class     - desktop         - re-named classes in commented code         - created "desktop-header - js     - tried to create media query assist in which after a certain window-width Header is removed and replaced with Desktop Header         (obviously this didn't work.)     - tried a different way to do same thing as above         (once again, didn't work.)
need to figure out to make it so the og. header is hidden when the desktop header is present and vice versa
also need to figure out how to get no gaps in on sale slider (this pisses me off. so much.)

once i've figured those out, i think i'll be able to complete the other media query problems(?)